requirements.txt
astroid==2.12.12
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
dill==0.3.6
idna==3.4
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
platformdirs==2.5.4
playsound==1.3.0
pycodestyle==2.9.1
PyExecJS==1.5.1
pylint==2.15.5
requests==2.28.1
requests-toolbelt==0.10.1
schedule==1.1.0
six==1.16.0
tomli==2.0.1
tomlkit==0.11.6
urllib3==1.26.12
wrapt==1.14.1
